rename services to app
