+++ src/app/determine-editor-version.ts
