+++ src/app/get-registry-auth.ts
