+++ src/cli/cmd-deps.ts
@@ -11,10 +11,10 @@ import { PackageUrl } from "../domain/package-url";
 import { unityRegistry } from "../domain/registry";
 import { SemanticVersion } from "../domain/semantic-version";
 import { DebugLog } from "../logging";
-import { ResolveDependencies } from "../services/dependency-resolving";
-import { GetLatestVersion } from "../services/get-latest-version";
-import { GetRegistryAuth } from "../services/get-registry-auth";
-import { ParseEnv } from "../services/parse-env";
+import { ResolveDependencies } from "../app/dependency-resolving";
+import { GetLatestVersion } from "../app/get-latest-version";
+import { GetRegistryAuth } from "../app/get-registry-auth";
+import { ParseEnv } from "../app/parse-env";
 import { queryAllRegistriesLazy } from "../utils/sources";
 import { isZod } from "../utils/zod-utils";
 import { stringifyDependencyGraph } from "./dependency-logging";
