+++ src/cli/cmd-search.ts
@@ -1,9 +1,9 @@
 import { Logger } from "npmlog";
 import * as os from "os";
 import { DebugLog } from "../logging";
-import { GetRegistryAuth } from "../services/get-registry-auth";
-import { ParseEnv } from "../services/parse-env";
-import { SearchPackages } from "../services/search-packages";
+import { GetRegistryAuth } from "../app/get-registry-auth";
+import { ParseEnv } from "../app/parse-env";
+import { SearchPackages } from "../app/search-packages";
 import { CmdOptions } from "./options";
 import { formatAsTable } from "./output-formatting";
 import { ResultCodes } from "./result-codes";
