+++ src/cli/index.ts
@@ -12,15 +12,15 @@ import { makeNpmRegistryClient } from "../io/reg-client";
 import { getHomePathFromEnv } from "../io/special-paths";
 import { getUpmConfigPath as getUpmConfigPathUsing } from "../io/upm-config-io";
 import { DebugLog } from "../logging";
-import { resolveDependencies as resolveDependenciesUsing } from "../services/dependency-resolving";
-import { determineEditorVersionUsing } from "../services/determine-editor-version";
-import { getLatestVersion as getLatestVersionUsing } from "../services/get-latest-version";
-import { getRegistryAuthUsing } from "../services/get-registry-auth";
-import { getRegistryPackumentVersionUsing } from "../services/get-registry-packument-version";
-import { login as loginUsing } from "../services/login";
-import { makeParseEnv } from "../services/parse-env";
-import { removePackages as removePackagesUsing } from "../services/remove-packages";
-import { searchPackagesUsing } from "../services/search-packages";
+import { resolveDependencies as resolveDependenciesUsing } from "../app/dependency-resolving";
+import { determineEditorVersionUsing } from "../app/determine-editor-version";
+import { getLatestVersion as getLatestVersionUsing } from "../app/get-latest-version";
+import { getRegistryAuthUsing } from "../app/get-registry-auth";
+import { getRegistryPackumentVersionUsing } from "../app/get-registry-packument-version";
+import { login as loginUsing } from "../app/login";
+import { makeParseEnv } from "../app/parse-env";
+import { removePackages as removePackagesUsing } from "../app/remove-packages";
+import { searchPackagesUsing } from "../app/search-packages";
 import { eachValue } from "./cli-parsing";
 import { makeAddCmd } from "./cmd-add";
 import { makeDepsCmd } from "./cmd-deps";
