+++ test/unit/app/get-registry-packument-version.test.ts
@@ -3,7 +3,7 @@ import { DomainName } from "../../../src/domain/domain-name";
 import { Registry } from "../../../src/domain/registry";
 import { SemanticVersion } from "../../../src/domain/semantic-version";
 import { GetRegistryPackument } from "../../../src/io/packument-io";
-import { FetchRegistryPackumentVersion } from "../../../src/services/get-registry-packument-version";
+import { FetchRegistryPackumentVersion } from "../../../src/app/get-registry-packument-version";
 import { buildPackument } from "../domain/data-packument";
 import { exampleRegistryUrl } from "../domain/data-registry";
 import { mockFunctionOfType } from "./func.mock";
