+++ test/unit/app/login.test.ts
@@ -1,8 +1,8 @@
 import { GetAuthToken } from "../../../src/io/get-auth-token";
 import { noopLogger } from "../../../src/logging";
-import { UpmConfigLogin } from "../../../src/services/login";
-import { StoreNpmAuthToken } from "../../../src/services/put-npm-auth-token";
-import { PutRegistryAuth } from "../../../src/services/put-registry-auth";
+import { UpmConfigLogin } from "../../../src/app/login";
+import { StoreNpmAuthToken } from "../../../src/app/put-npm-auth-token";
+import { PutRegistryAuth } from "../../../src/app/put-registry-auth";
 import { exampleRegistryUrl } from "../domain/data-registry";
 import { mockFunctionOfType } from "./func.mock";
